+++ .vscode/settings.json
@@ -50,5 +50,6 @@
         "xtr1common": "cpp",
         "xtree": "cpp",
         "xutility": "cpp"
-    }
+    },
+    "C_Cpp.errorSquiggles": "Enabled"
 }
