+++ headerFiles/global.hh
@@ -1,0 +1,14 @@
+#ifndef GLOBAL_HMS
+#define GLOBAL_HMS
+
+#include <string>
+using namespace std;
+
+#ifndef DATE_DEFED
+#define DATE_DEFED
+int date;
+#endif // !DATE_DEFED
+int power(int n = 1, int exp = 0);
+int strToNum(string s);
+
+#endif // !GLOBAL_HMS
